pyproject.toml
[tool.poetry]
name = "django-reddit-challenge"
version = "0.1.0"
description = ""
authors = ["jlucasoliveira"]

[tool.poetry.dependencies]
python = "^3.7.10"
Django = "3.0.7"
django-dotenv = "1.4.2"
s3-environ = "0.0.5"
djangorestframework = "3.11.0"
django-filter = "2.2.0"
drf-nested-routers = "0.91.0"
dj-database-url = "0.5.0"
django-storages = "1.8.0"
gunicorn = "20.0.4"
whitenoise = "5.0.1"
django-cors-headers = "3.2.1"
django-rest-auth = "0.9.2"
django-amazon-ses = "3.0.2"
ipython = "7.12.0"
ipython_genutils = "0.2.0"
requests = "^2.25.1"
sentry-sdk = "^1.3.0"
django-allauth = "^0.44.0"
psycopg2-binary = "=2.8.6"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
include = '\.pyi?$'
exclude = '''
(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)
'''